SWITCHYARD-1216 create separate feature for bpmn2
