+++ eclipse/features/org.switchyard.tools.bpmn2.feature/build.properties
@@ -1,5 +1,4 @@
 bin.includes = feature.xml,\
                feature.properties,\
-               license.html,\
-               p2.inf
+               license.html
 src.includes = license.html
