+++ eclipse/features/org.switchyard.tools.bpmn2.feature/feature.properties
@@ -1,9 +1,9 @@
-# properties file for org.switchyard.tools.feature
-featureName=SwitchYard
+# properties file for org.switchyard.tools.sca.feature
+featureName=SwitchYard BPM Support
 featureProvider=www.jboss.org/switchyard
 
 # "description" property - description of the feature
-description=Tooling for building SwitchYard applications.
+description=Provides capabilities for working with BPMN2 components in SwitchYard applications.
 
 # "licenseURL" property - URL of the "Feature License"
 # do not translate value - just change to point to a locale-specific HTML page
