+++ eclipse/features/org.switchyard.tools.bpmn2.feature/feature.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <feature
-      id="org.switchyard.tools.bpel.feature"
+      id="org.switchyard.tools.bpmn2.feature"
       label="%featureName"
       version="0.8.0.qualifier"
       provider-name="%featureProvider">
@@ -19,10 +19,11 @@
 
    <requires>
       <import feature="org.switchyard.tools.feature" version="0.8.0.qualifier"/>
+      <import plugin="org.eclipse.bpmn2.modeler.ui" version="0.2.2" match="equivalent"/>
    </requires>
 
    <plugin
-         id="org.switchyard.tools.ui.bpel"
+         id="org.switchyard.tools.ui.bpmn2"
          download-size="0"
          install-size="0"
          version="0.0.0"
