+++ eclipse/features/org.switchyard.tools.bpmn2.feature/pom.xml
@@ -1,0 +1,15 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+  <modelVersion>4.0.0</modelVersion>
+  <parent>
+    <groupId>org.switchyard.tools</groupId>
+    <artifactId>switchyard-tools-eclipse-features</artifactId>
+    <version>0.8.0-SNAPSHOT</version>
+    <relativePath>../pom.xml</relativePath>
+  </parent>
+  <artifactId>org.switchyard.tools.bpmn2.feature</artifactId>
+  <packaging>eclipse-feature</packaging>
+  <name>SwitchYard: Eclipse SwitchYard BPMN2 Feature</name>
+  <description>Eclipse SwitchYard BPMN2 component support.</description>
+
+</project>
