+++ eclipse/features/org.switchyard.tools.editor.feature/feature.xml
@@ -17,25 +17,4 @@
       %license
    </license>
 
-   <requires>
-      <import feature="org.switchyard.tools.feature" version="0.8.0.qualifier" match="greaterOrEqual"/>
-      <import feature="org.eclipse.graphiti.feature" version="0.9.0" match="equivalent"/>
-      <import feature="org.jboss.ide.eclipse.freemarker.feature" version="1.2.0" match="greaterOrEqual"/>
-      <import feature="org.eclipse.emf.compare" version="1.2.0" match="compatible"/>
-   </requires>
-
-   <plugin
-         id="org.switchyard.tools.ui.editor"
-         download-size="0"
-         install-size="0"
-         version="0.0.0"
-         unpack="false"/>
-
-   <plugin
-         id="org.switchyard.tools.ui.bpmn2"
-         download-size="0"
-         install-size="0"
-         version="0.0.0"
-         unpack="false"/>
-
 </feature>
