+++ eclipse/plugins/org.switchyard.tools.ui.bpmn2/META-INF/MANIFEST.MF
@@ -1,6 +1,6 @@
 Manifest-Version: 1.0
 Bundle-ManifestVersion: 2
-Bundle-Name: SwitchYard Tools BPMN2 Editor Extensions
+Bundle-Name: SwitchYard Tools BPMN2 Support
 Bundle-SymbolicName: org.switchyard.tools.ui.bpmn2;singleton:=true
 Bundle-Version: 0.8.0.qualifier
 Bundle-Vendor: www.jboss.org
@@ -9,6 +9,8 @@ Bundle-ActivationPolicy: lazy
 Require-Bundle: org.eclipse.ui,
  org.eclipse.core.runtime,
  org.eclipse.core.resources,
+ org.eclipse.jdt.core,
+ org.eclipse.ui.ide,
  org.eclipse.ui.views.properties.tabbed,
  org.switchyard.tools.ui,
  org.switchyard.tools.ui.editor,
@@ -16,10 +18,11 @@ Require-Bundle: org.eclipse.ui,
  org.eclipse.graphiti,
  org.eclipse.graphiti.ui,
  org.eclipse.gef,
+ org.eclipse.emf.edit.ui,
  org.eclipse.emf.transaction,
- org.eclipse.bpmn2;resolution:=optional,
- org.eclipse.bpmn2.modeler.core;bundle-version="[0.2.0,0.2.2]";resolution:=optional,
- org.eclipse.bpmn2.modeler.ui;bundle-version="[0.2.0,0.2.2]";resolution:=optional,
- org.eclipse.bpmn2.modeler.runtime.jboss.jbpm5;bundle-version="[0.2.0,0.2.2]";resolution:=optional
+ org.eclipse.bpmn2,
+ org.eclipse.bpmn2.modeler.core;bundle-version="[0.2.0,0.3.0)",
+ org.eclipse.bpmn2.modeler.ui;bundle-version="[0.2.0,0.3.0)",
+ org.eclipse.bpmn2.modeler.runtime.jboss.jbpm5;bundle-version="[0.2.0,0.3.0)"
 Bundle-ClassPath: .
 Bundle-Activator: org.switchyard.tools.ui.bpmn2.Activator
