+++ eclipse/plugins/org.switchyard.tools.ui.bpmn2/src/org/switchyard/tools/ui/bpmn2/component/BPMComponentTypeExtension.java
@@ -9,7 +9,7 @@
  * Red Hat, Inc. - initial API and implementation 
  *
  ******************************************************************************/
-package org.switchyard.tools.ui.editor.components.bpm;
+package org.switchyard.tools.ui.bpmn2.component;
 
 import java.util.Collections;
 import java.util.List;
