+++ eclipse/plugins/org.switchyard.tools.ui.bpmn2/src/org/switchyard/tools/ui/bpmn2/component/BPMLoggerTable.java
@@ -10,13 +10,16 @@
  *
  * @author bfitzpat
  ******************************************************************************/
-package org.switchyard.tools.ui.editor.components.bpm;
+package org.switchyard.tools.ui.bpmn2.component;
+
+import java.math.BigInteger;
 
 import javax.swing.event.ChangeEvent;
 import javax.swing.event.ChangeListener;
 
 import org.eclipse.core.runtime.ListenerList;
 import org.eclipse.emf.ecore.EObject;
+import org.eclipse.emf.edit.ui.provider.PropertyDescriptor.EDataTypeCellEditor;
 import org.eclipse.emf.transaction.RecordingCommand;
 import org.eclipse.emf.transaction.TransactionalEditingDomain;
 import org.eclipse.jface.viewers.CellEditor;
@@ -40,10 +43,11 @@ import org.eclipse.swt.widgets.Button;
 import org.eclipse.swt.widgets.Composite;
 import org.eclipse.swt.widgets.TableColumn;
 import org.eclipse.swt.widgets.TableItem;
-import org.switchyard.tools.models.switchyard1_0.bpm.ActionType;
-import org.switchyard.tools.models.switchyard1_0.bpm.ActionType1;
 import org.switchyard.tools.models.switchyard1_0.bpm.BPMFactory;
 import org.switchyard.tools.models.switchyard1_0.bpm.BPMImplementationType;
+import org.switchyard.tools.models.switchyard1_0.bpm.BPMPackage;
+import org.switchyard.tools.models.switchyard1_0.bpm.LoggerType;
+import org.switchyard.tools.models.switchyard1_0.bpm.LoggerType1;
 import org.switchyard.tools.ui.editor.diagram.shared.TableColumnLayout;
 import org.switchyard.tools.ui.editor.impl.SwitchyardSCAEditor;
 
@@ -51,7 +55,7 @@ import org.switchyard.tools.ui.editor.impl.SwitchyardSCAEditor;
  * @author bfitzpat
  * 
  */
-public class BPMActionTable extends Composite implements ICellModifier {
+public class BPMLoggerTable extends Composite implements ICellModifier {
 
     private class PropertyTreeContentProvider implements IStructuredContentProvider {
 
@@ -67,8 +71,8 @@ public class BPMActionTable extends Composite implements ICellModifier {
         public Object[] getElements(Object inputElement) {
             if (inputElement instanceof BPMImplementationType) {
                 final BPMImplementationType bpmImpl = (BPMImplementationType) inputElement;
-                if (bpmImpl.getActions() != null) {
-                    return bpmImpl.getActions().getAction().toArray();
+                if (bpmImpl.getLoggers() != null) {
+                    return bpmImpl.getLoggers().getLogger().toArray();
                 }
             }
             return new Object[0];
@@ -86,9 +90,9 @@ public class BPMActionTable extends Composite implements ICellModifier {
 
         @Override
         public boolean isLabelProperty(Object element, String property) {
-            if (element instanceof ActionType1 && property.equalsIgnoreCase(VALUE_COLUMN)) {
+            if (element instanceof LoggerType1 && property.equalsIgnoreCase(LOG_COLUMN)) {
                 return true;
-            } else if (element instanceof ActionType1 && property.equalsIgnoreCase(ENTRY_POINT_COLUMN)) {
+            } else if (element instanceof LoggerType1 && property.equalsIgnoreCase(TYPE_COLUMN)) {
                 return true;
             }
             return false;
@@ -105,12 +109,20 @@ public class BPMActionTable extends Composite implements ICellModifier {
 
         @Override
         public String getColumnText(Object element, int columnIndex) {
-            if (element instanceof ActionType1 && columnIndex == 0) {
-                ActionType1 tp = (ActionType1) element;
-                return (String) tp.getType().getLiteral();
-            } else if (element instanceof ActionType1 && columnIndex == 1) {
-                ActionType1 tp = (ActionType1) element;
-                return tp.getOperation();
+            if (element instanceof LoggerType1 && columnIndex == 0) {
+                if (((LoggerType1) element).getLog() != null) {
+                    return ((LoggerType1) element).getLog();
+                } else {
+                    return "";
+                }
+            } else if (element instanceof LoggerType1 && columnIndex == 1) {
+                return ((LoggerType1) element).getType().getLiteral();
+            } else if (element instanceof LoggerType1 && columnIndex == 2) {
+                if (((LoggerType1) element).getInterval() != null) {
+                    return ((LoggerType1) element).getInterval().toString();
+                } else {
+                    return "";
+                }
             }
             return null;
         }
@@ -118,16 +130,11 @@ public class BPMActionTable extends Composite implements ICellModifier {
 
     private TableViewer _propertyTreeTable;
 
-    /**
-     * Value column.
-     */
-    public static final String VALUE_COLUMN = "value";
-    /**
-     * Entry point column.
-     */
-    public static final String ENTRY_POINT_COLUMN = "entryPoint";
+    private static final String LOG_COLUMN = "log";
+    private static final String TYPE_COLUMN = "type";
+    private static final String INTERVAL_COLUMN = "interval";
 
-    private static final String[] TREE_COLUMNS = new String[] {VALUE_COLUMN, ENTRY_POINT_COLUMN };
+    private static final String[] TREE_COLUMNS = new String[] {LOG_COLUMN, TYPE_COLUMN, INTERVAL_COLUMN };
 
     private Button _mAddButton;
     private Button _mRemoveButton;
@@ -142,7 +149,7 @@ public class BPMActionTable extends Composite implements ICellModifier {
      * @param parent Composite parent
      * @param style any SWT style bits to pass along
      */
-    public BPMActionTable(Composite parent, int style) {
+    public BPMLoggerTable(Composite parent, int style) {
         this(parent, style, false);
     }
 
@@ -153,7 +160,7 @@ public class BPMActionTable extends Composite implements ICellModifier {
      * @param style any SWT style bits
      * @param isReadOnly boolean flag
      */
-    public BPMActionTable(Composite parent, int style, boolean isReadOnly) {
+    public BPMLoggerTable(Composite parent, int style, boolean isReadOnly) {
         super(parent, style);
         this._isReadOnly = isReadOnly;
         this._changeListeners = new ListenerList();
@@ -170,7 +177,7 @@ public class BPMActionTable extends Composite implements ICellModifier {
         setLayout(gridLayout);
 
         Composite tableComposite = new Composite(this, additionalStyles);
-        GridData gd11 = new GridData(SWT.FILL, SWT.FILL, true, true, 1, 2);
+        GridData gd11 = new GridData(SWT.FILL, SWT.FILL, true, false, 1, 2);
         gd11.heightHint = 100;
         tableComposite.setLayoutData(gd11);
 
@@ -181,12 +188,15 @@ public class BPMActionTable extends Composite implements ICellModifier {
         TableColumnLayout tableLayout = new TableColumnLayout();
         tableComposite.setLayout(tableLayout);
 
-        TableColumn valueColumn = new TableColumn(_propertyTreeTable.getTable(), SWT.LEFT);
-        valueColumn.setText("Type");
-        tableLayout.setColumnData(valueColumn, new ColumnWeightData(100, 150, true));
-        TableColumn entryPointColumn = new TableColumn(_propertyTreeTable.getTable(), SWT.LEFT);
-        entryPointColumn.setText("Operation");
-        tableLayout.setColumnData(entryPointColumn, new ColumnWeightData(100, 150, true));
+        TableColumn logColumn = new TableColumn(_propertyTreeTable.getTable(), SWT.LEFT);
+        logColumn.setText("Log");
+        tableLayout.setColumnData(logColumn, new ColumnWeightData(100, 150, true));
+        TableColumn typeColumn = new TableColumn(_propertyTreeTable.getTable(), SWT.LEFT);
+        typeColumn.setText("Type");
+        tableLayout.setColumnData(typeColumn, new ColumnWeightData(100, 150, true));
+        TableColumn intervalColumn = new TableColumn(_propertyTreeTable.getTable(), SWT.LEFT);
+        intervalColumn.setText("Interval");
+        tableLayout.setColumnData(intervalColumn, new ColumnWeightData(100, 150, true));
 
         _propertyTreeTable.setColumnProperties(TREE_COLUMNS);
 
@@ -196,10 +206,11 @@ public class BPMActionTable extends Composite implements ICellModifier {
 
         _propertyTreeTable.setCellModifier(this);
         _propertyTreeTable.setCellEditors(new CellEditor[] {
-                new ComboBoxCellEditor(_propertyTreeTable.getTable(), new String[] {
-                        ActionType.STARTPROCESS.getLiteral(), ActionType.SIGNALEVENT.getLiteral(),
-                        ActionType.ABORTPROCESSINSTANCE.getLiteral() }),
-                new TextCellEditor(_propertyTreeTable.getTable()) });
+                new TextCellEditor(_propertyTreeTable.getTable()),
+                new ComboBoxCellEditor(_propertyTreeTable.getTable(), new String[] {LoggerType.CONSOLE.getLiteral(),
+                        LoggerType.FILE.getLiteral(), LoggerType.THREADEDFILE.getLiteral() }),
+                new EDataTypeCellEditor(BPMPackage.eINSTANCE.getLoggerType1_Interval().getEAttributeType(),
+                        _propertyTreeTable.getTable()) });
 
         _mAddButton = new Button(this, SWT.NONE);
         _mAddButton.setLayoutData(new GridData(SWT.FILL, SWT.BEGINNING, false, false));
@@ -257,29 +268,25 @@ public class BPMActionTable extends Composite implements ICellModifier {
     protected void addPropertyToList() {
         if (getTargetObject() instanceof BPMImplementationType) {
             final BPMImplementationType impl = (BPMImplementationType) getTargetObject();
+            final LoggerType1 newLogger = BPMFactory.eINSTANCE.createLoggerType1();
+            newLogger.setLog("NewLogger");
             if (impl.eContainer() != null) {
                 TransactionalEditingDomain domain = SwitchyardSCAEditor.getActiveEditor().getEditingDomain();
                 domain.getCommandStack().execute(new RecordingCommand(domain) {
                     @Override
                     protected void doExecute() {
-                        ActionType1 newAction = BPMFactory.eINSTANCE.createActionType1();
-                        newAction.setType(ActionType.STARTPROCESS);
-                        newAction.setOperation("NewAction");
-                        if (impl.getActions() == null) {
-                            impl.setActions(BPMFactory.eINSTANCE.createActionsType());
+                        if (impl.getLoggers() == null) {
+                            impl.setLoggers(BPMFactory.eINSTANCE.createLoggersType());
                         }
-                        impl.getActions().getAction().add(newAction);
+                        impl.getLoggers().getLogger().add(newLogger);
                         getTableViewer().refresh(true);
                     }
                 });
             } else {
-                ActionType1 newAction = BPMFactory.eINSTANCE.createActionType1();
-                newAction.setType(ActionType.STARTPROCESS);
-                newAction.setOperation("NewAction");
-                if (impl.getActions() == null) {
-                    impl.setActions(BPMFactory.eINSTANCE.createActionsType());
+                if (impl.getLoggers() == null) {
+                    impl.setLoggers(BPMFactory.eINSTANCE.createLoggersType());
                 }
-                impl.getActions().getAction().add(newAction);
+                impl.getLoggers().getLogger().add(newLogger);
                 getTableViewer().refresh(true);
             }
             fireChangedEvent(this);
@@ -292,23 +299,23 @@ public class BPMActionTable extends Composite implements ICellModifier {
     protected void removeFromList() {
         if (getTargetObject() instanceof BPMImplementationType) {
             final BPMImplementationType impl = (BPMImplementationType) getTargetObject();
-            final ActionType1 actionToRemove = getTableSelection();
+            final LoggerType1 actionToRemove = getTableSelection();
             if (impl.eContainer() != null) {
                 TransactionalEditingDomain domain = SwitchyardSCAEditor.getActiveEditor().getEditingDomain();
                 domain.getCommandStack().execute(new RecordingCommand(domain) {
                     @Override
                     protected void doExecute() {
-                        impl.getActions().getAction().remove(actionToRemove);
-                        if (impl.getActions().getAction().isEmpty()) {
-                            impl.setActions(null);
+                        impl.getLoggers().getLogger().remove(actionToRemove);
+                        if (impl.getLoggers().getLogger().isEmpty()) {
+                            impl.setLoggers(null);
                         }
                         getTableViewer().refresh(true);
                     }
                 });
             } else {
-                impl.getActions().getAction().remove(actionToRemove);
-                if (impl.getActions().getAction().isEmpty()) {
-                    impl.setActions(null);
+                impl.getLoggers().getLogger().remove(actionToRemove);
+                if (impl.getLoggers().getLogger().isEmpty()) {
+                    impl.setLoggers(null);
                 }
                 getTableViewer().refresh(true);
             }
@@ -316,11 +323,11 @@ public class BPMActionTable extends Composite implements ICellModifier {
         }
     }
 
-    protected ActionType1 getTableSelection() {
+    protected LoggerType1 getTableSelection() {
         if (_propertyTreeTable != null && !_propertyTreeTable.getSelection().isEmpty()) {
             IStructuredSelection ssel = (IStructuredSelection) _propertyTreeTable.getSelection();
-            if (ssel.getFirstElement() instanceof ActionType1) {
-                return (ActionType1) ssel.getFirstElement();
+            if (ssel.getFirstElement() instanceof LoggerType1) {
+                return (LoggerType1) ssel.getFirstElement();
             }
         }
         return null;
@@ -413,14 +420,16 @@ public class BPMActionTable extends Composite implements ICellModifier {
      *      java.lang.String)
      */
     public Object getValue(Object element, String property) {
-        if (element instanceof ActionType1 && property.equalsIgnoreCase(VALUE_COLUMN)) {
-            return new Integer(((ActionType1) element).getType().getValue());
-        } else if (element instanceof ActionType1 && property.equalsIgnoreCase(ENTRY_POINT_COLUMN)) {
-            if (((ActionType1) element).getOperation() != null) {
-                return ((ActionType1) element).getOperation();
+        if (element instanceof LoggerType1 && property.equalsIgnoreCase(LOG_COLUMN)) {
+            if (((LoggerType1) element).getLog() != null) {
+                return ((LoggerType1) element).getLog();
             } else {
                 return "";
             }
+        } else if (element instanceof LoggerType1 && property.equalsIgnoreCase(TYPE_COLUMN)) {
+            return new Integer(((LoggerType1) element).getType().getValue());
+        } else if (element instanceof LoggerType1 && property.equalsIgnoreCase(INTERVAL_COLUMN)) {
+            return ((LoggerType1) element).getInterval();
         }
         return null;
     }
@@ -434,7 +443,31 @@ public class BPMActionTable extends Composite implements ICellModifier {
      *      java.lang.String, java.lang.Object)
      */
     public void modify(Object element, String property, final Object value) {
-        if (element instanceof TableItem && property.equalsIgnoreCase(VALUE_COLUMN)) {
+        if (element instanceof TableItem && property.equalsIgnoreCase(LOG_COLUMN)) {
+            final TableItem ti = (TableItem) element;
+            if (getTargetObject() instanceof BPMImplementationType) {
+                final BPMImplementationType impl = (BPMImplementationType) getTargetObject();
+                final LoggerType1 parm = (LoggerType1) ti.getData();
+                if ((value == null && parm.getLog() == null) || (value != null && value.equals(parm.getLog()))) {
+                    return;
+                }
+                if (impl.eContainer() != null) {
+                    TransactionalEditingDomain domain = SwitchyardSCAEditor.getActiveEditor().getEditingDomain();
+                    domain.getCommandStack().execute(new RecordingCommand(domain) {
+                        @Override
+                        protected void doExecute() {
+                            parm.setLog((String) value);
+                            getTableViewer().refresh(true);
+                        }
+                    });
+                } else {
+                    parm.setLog((String) value);
+                    getTableViewer().refresh(true);
+                }
+            }
+            fireChangedEvent(this);
+            // validate();
+        } else if (element instanceof TableItem && property.equalsIgnoreCase(TYPE_COLUMN)) {
             final TableItem ti = (TableItem) element;
             if (getTargetObject() instanceof BPMImplementationType) {
                 final BPMImplementationType impl = (BPMImplementationType) getTargetObject();
@@ -443,39 +476,38 @@ public class BPMActionTable extends Composite implements ICellModifier {
                     domain.getCommandStack().execute(new RecordingCommand(domain) {
                         @Override
                         protected void doExecute() {
-                            ActionType1 parm = (ActionType1) ti.getData();
-                            ActionType atype = ActionType.get(((Integer) value).intValue());
+                            LoggerType1 parm = (LoggerType1) ti.getData();
+                            LoggerType atype = LoggerType.get(((Integer) value).intValue());
                             parm.setType(atype);
                             getTableViewer().refresh(true);
                         }
                     });
                 } else {
-                    ActionType1 parm = (ActionType1) ti.getData();
-                    ActionType atype = ActionType.get(((Integer) value).intValue());
+                    LoggerType1 parm = (LoggerType1) ti.getData();
+                    LoggerType atype = LoggerType.get(((Integer) value).intValue());
                     parm.setType(atype);
                     getTableViewer().refresh(true);
                 }
             }
             fireChangedEvent(this);
             // validate();
-        } else if (element instanceof TableItem && property.equalsIgnoreCase(ENTRY_POINT_COLUMN)) {
+        } else if (element instanceof TableItem && property.equalsIgnoreCase(INTERVAL_COLUMN)) {
             final TableItem ti = (TableItem) element;
             if (getTargetObject() instanceof BPMImplementationType) {
                 final BPMImplementationType impl = (BPMImplementationType) getTargetObject();
-                final String newValue = value == null || ((String) value).length() == 0 ? null : (String) value;
                 if (impl.eContainer() != null) {
                     TransactionalEditingDomain domain = SwitchyardSCAEditor.getActiveEditor().getEditingDomain();
                     domain.getCommandStack().execute(new RecordingCommand(domain) {
                         @Override
                         protected void doExecute() {
-                            ActionType1 parm = (ActionType1) ti.getData();
-                            parm.setOperation(newValue);
+                            LoggerType1 parm = (LoggerType1) ti.getData();
+                            parm.setInterval((BigInteger) value);
                             getTableViewer().refresh(true);
                         }
                     });
                 } else {
-                    ActionType1 parm = (ActionType1) ti.getData();
-                    parm.setOperation(newValue);
+                    LoggerType1 parm = (LoggerType1) ti.getData();
+                    parm.setInterval((BigInteger) value);
                     getTableViewer().refresh(true);
                 }
             }
