+++ eclipse/plugins/org.switchyard.tools.ui.bpmn2/src/org/switchyard/tools/ui/bpmn2/component/BPMPropertyTable.java
@@ -10,21 +10,17 @@
  *
  * @author bfitzpat
  ******************************************************************************/
-package org.switchyard.tools.ui.editor.components.bpm;
-
-import java.math.BigInteger;
+package org.switchyard.tools.ui.bpmn2.component;
 
 import javax.swing.event.ChangeEvent;
 import javax.swing.event.ChangeListener;
 
 import org.eclipse.core.runtime.ListenerList;
 import org.eclipse.emf.ecore.EObject;
-import org.eclipse.emf.edit.ui.provider.PropertyDescriptor.EDataTypeCellEditor;
 import org.eclipse.emf.transaction.RecordingCommand;
 import org.eclipse.emf.transaction.TransactionalEditingDomain;
 import org.eclipse.jface.viewers.CellEditor;
 import org.eclipse.jface.viewers.ColumnWeightData;
-import org.eclipse.jface.viewers.ComboBoxCellEditor;
 import org.eclipse.jface.viewers.ICellModifier;
 import org.eclipse.jface.viewers.ILabelProviderListener;
 import org.eclipse.jface.viewers.IStructuredContentProvider;
@@ -45,9 +41,7 @@ import org.eclipse.swt.widgets.TableColumn;
 import org.eclipse.swt.widgets.TableItem;
 import org.switchyard.tools.models.switchyard1_0.bpm.BPMFactory;
 import org.switchyard.tools.models.switchyard1_0.bpm.BPMImplementationType;
-import org.switchyard.tools.models.switchyard1_0.bpm.BPMPackage;
-import org.switchyard.tools.models.switchyard1_0.bpm.LoggerType;
-import org.switchyard.tools.models.switchyard1_0.bpm.LoggerType1;
+import org.switchyard.tools.models.switchyard1_0.bpm.PropertyType;
 import org.switchyard.tools.ui.editor.diagram.shared.TableColumnLayout;
 import org.switchyard.tools.ui.editor.impl.SwitchyardSCAEditor;
 
@@ -55,7 +49,7 @@ import org.switchyard.tools.ui.editor.impl.SwitchyardSCAEditor;
  * @author bfitzpat
  * 
  */
-public class BPMLoggerTable extends Composite implements ICellModifier {
+public class BPMPropertyTable extends Composite implements ICellModifier {
 
     private class PropertyTreeContentProvider implements IStructuredContentProvider {
 
@@ -71,8 +65,8 @@ public class BPMLoggerTable extends Composite implements ICellModifier {
         public Object[] getElements(Object inputElement) {
             if (inputElement instanceof BPMImplementationType) {
                 final BPMImplementationType bpmImpl = (BPMImplementationType) inputElement;
-                if (bpmImpl.getLoggers() != null) {
-                    return bpmImpl.getLoggers().getLogger().toArray();
+                if (bpmImpl.getProperties() != null) {
+                    return bpmImpl.getProperties().getProperty().toArray();
                 }
             }
             return new Object[0];
@@ -90,9 +84,9 @@ public class BPMLoggerTable extends Composite implements ICellModifier {
 
         @Override
         public boolean isLabelProperty(Object element, String property) {
-            if (element instanceof LoggerType1 && property.equalsIgnoreCase(LOG_COLUMN)) {
+            if (element instanceof PropertyType && property.equalsIgnoreCase(NAME_COLUMN)) {
                 return true;
-            } else if (element instanceof LoggerType1 && property.equalsIgnoreCase(TYPE_COLUMN)) {
+            } else if (element instanceof PropertyType && property.equalsIgnoreCase(VALUE_COLUMN)) {
                 return true;
             }
             return false;
@@ -109,20 +103,10 @@ public class BPMLoggerTable extends Composite implements ICellModifier {
 
         @Override
         public String getColumnText(Object element, int columnIndex) {
-            if (element instanceof LoggerType1 && columnIndex == 0) {
-                if (((LoggerType1) element).getLog() != null) {
-                    return ((LoggerType1) element).getLog();
-                } else {
-                    return "";
-                }
-            } else if (element instanceof LoggerType1 && columnIndex == 1) {
-                return ((LoggerType1) element).getType().getLiteral();
-            } else if (element instanceof LoggerType1 && columnIndex == 2) {
-                if (((LoggerType1) element).getInterval() != null) {
-                    return ((LoggerType1) element).getInterval().toString();
-                } else {
-                    return "";
-                }
+            if (element instanceof PropertyType && columnIndex == 0) {
+                return ((PropertyType) element).getName();
+            } else if (element instanceof PropertyType && columnIndex == 1) {
+                return ((PropertyType) element).getValue();
             }
             return null;
         }
@@ -130,11 +114,16 @@ public class BPMLoggerTable extends Composite implements ICellModifier {
 
     private TableViewer _propertyTreeTable;
 
-    private static final String LOG_COLUMN = "log";
-    private static final String TYPE_COLUMN = "type";
-    private static final String INTERVAL_COLUMN = "interval";
+    /**
+     * Value column.
+     */
+    public static final String NAME_COLUMN = "name";
+    /**
+     * Entry point column.
+     */
+    public static final String VALUE_COLUMN = "value";
 
-    private static final String[] TREE_COLUMNS = new String[] {LOG_COLUMN, TYPE_COLUMN, INTERVAL_COLUMN };
+    private static final String[] TREE_COLUMNS = new String[] {NAME_COLUMN, VALUE_COLUMN };
 
     private Button _mAddButton;
     private Button _mRemoveButton;
@@ -149,7 +138,7 @@ public class BPMLoggerTable extends Composite implements ICellModifier {
      * @param parent Composite parent
      * @param style any SWT style bits to pass along
      */
-    public BPMLoggerTable(Composite parent, int style) {
+    public BPMPropertyTable(Composite parent, int style) {
         this(parent, style, false);
     }
 
@@ -160,7 +149,7 @@ public class BPMLoggerTable extends Composite implements ICellModifier {
      * @param style any SWT style bits
      * @param isReadOnly boolean flag
      */
-    public BPMLoggerTable(Composite parent, int style, boolean isReadOnly) {
+    public BPMPropertyTable(Composite parent, int style, boolean isReadOnly) {
         super(parent, style);
         this._isReadOnly = isReadOnly;
         this._changeListeners = new ListenerList();
@@ -188,15 +177,12 @@ public class BPMLoggerTable extends Composite implements ICellModifier {
         TableColumnLayout tableLayout = new TableColumnLayout();
         tableComposite.setLayout(tableLayout);
 
-        TableColumn logColumn = new TableColumn(_propertyTreeTable.getTable(), SWT.LEFT);
-        logColumn.setText("Log");
-        tableLayout.setColumnData(logColumn, new ColumnWeightData(100, 150, true));
-        TableColumn typeColumn = new TableColumn(_propertyTreeTable.getTable(), SWT.LEFT);
-        typeColumn.setText("Type");
-        tableLayout.setColumnData(typeColumn, new ColumnWeightData(100, 150, true));
-        TableColumn intervalColumn = new TableColumn(_propertyTreeTable.getTable(), SWT.LEFT);
-        intervalColumn.setText("Interval");
-        tableLayout.setColumnData(intervalColumn, new ColumnWeightData(100, 150, true));
+        TableColumn nameColumn = new TableColumn(_propertyTreeTable.getTable(), SWT.LEFT);
+        nameColumn.setText("Name");
+        tableLayout.setColumnData(nameColumn, new ColumnWeightData(100, 150, true));
+        TableColumn valueColumn = new TableColumn(_propertyTreeTable.getTable(), SWT.LEFT);
+        valueColumn.setText("Value");
+        tableLayout.setColumnData(valueColumn, new ColumnWeightData(100, 150, true));
 
         _propertyTreeTable.setColumnProperties(TREE_COLUMNS);
 
@@ -205,12 +191,8 @@ public class BPMLoggerTable extends Composite implements ICellModifier {
         _propertyTreeTable.setContentProvider(new PropertyTreeContentProvider());
 
         _propertyTreeTable.setCellModifier(this);
-        _propertyTreeTable.setCellEditors(new CellEditor[] {
-                new TextCellEditor(_propertyTreeTable.getTable()),
-                new ComboBoxCellEditor(_propertyTreeTable.getTable(), new String[] {LoggerType.CONSOLE.getLiteral(),
-                        LoggerType.FILE.getLiteral(), LoggerType.THREADEDFILE.getLiteral() }),
-                new EDataTypeCellEditor(BPMPackage.eINSTANCE.getLoggerType1_Interval().getEAttributeType(),
-                        _propertyTreeTable.getTable()) });
+        _propertyTreeTable.setCellEditors(new CellEditor[] {new TextCellEditor(_propertyTreeTable.getTable()),
+                new TextCellEditor(_propertyTreeTable.getTable()) });
 
         _mAddButton = new Button(this, SWT.NONE);
         _mAddButton.setLayoutData(new GridData(SWT.FILL, SWT.BEGINNING, false, false));
@@ -268,25 +250,26 @@ public class BPMLoggerTable extends Composite implements ICellModifier {
     protected void addPropertyToList() {
         if (getTargetObject() instanceof BPMImplementationType) {
             final BPMImplementationType impl = (BPMImplementationType) getTargetObject();
-            final LoggerType1 newLogger = BPMFactory.eINSTANCE.createLoggerType1();
-            newLogger.setLog("NewLogger");
+            final PropertyType newAction = BPMFactory.eINSTANCE.createPropertyType();
+            newAction.setName("property");
+            newAction.setValue("value");
             if (impl.eContainer() != null) {
                 TransactionalEditingDomain domain = SwitchyardSCAEditor.getActiveEditor().getEditingDomain();
                 domain.getCommandStack().execute(new RecordingCommand(domain) {
                     @Override
                     protected void doExecute() {
-                        if (impl.getLoggers() == null) {
-                            impl.setLoggers(BPMFactory.eINSTANCE.createLoggersType());
+                        if (impl.getProperties() == null) {
+                            impl.setProperties(BPMFactory.eINSTANCE.createPropertiesType());
                         }
-                        impl.getLoggers().getLogger().add(newLogger);
+                        impl.getProperties().getProperty().add(newAction);
                         getTableViewer().refresh(true);
                     }
                 });
             } else {
-                if (impl.getLoggers() == null) {
-                    impl.setLoggers(BPMFactory.eINSTANCE.createLoggersType());
+                if (impl.getProperties() == null) {
+                    impl.setProperties(BPMFactory.eINSTANCE.createPropertiesType());
                 }
-                impl.getLoggers().getLogger().add(newLogger);
+                impl.getProperties().getProperty().add(newAction);
                 getTableViewer().refresh(true);
             }
             fireChangedEvent(this);
@@ -299,23 +282,23 @@ public class BPMLoggerTable extends Composite implements ICellModifier {
     protected void removeFromList() {
         if (getTargetObject() instanceof BPMImplementationType) {
             final BPMImplementationType impl = (BPMImplementationType) getTargetObject();
-            final LoggerType1 actionToRemove = getTableSelection();
+            final PropertyType actionToRemove = getTableSelection();
             if (impl.eContainer() != null) {
                 TransactionalEditingDomain domain = SwitchyardSCAEditor.getActiveEditor().getEditingDomain();
                 domain.getCommandStack().execute(new RecordingCommand(domain) {
                     @Override
                     protected void doExecute() {
-                        impl.getLoggers().getLogger().remove(actionToRemove);
-                        if (impl.getLoggers().getLogger().isEmpty()) {
-                            impl.setLoggers(null);
+                        impl.getProperties().getProperty().remove(actionToRemove);
+                        if (impl.getProperties().getProperty().isEmpty()) {
+                            impl.setProperties(null);
                         }
                         getTableViewer().refresh(true);
                     }
                 });
             } else {
-                impl.getLoggers().getLogger().remove(actionToRemove);
-                if (impl.getLoggers().getLogger().isEmpty()) {
-                    impl.setLoggers(null);
+                impl.getProperties().getProperty().remove(actionToRemove);
+                if (impl.getProperties().getProperty().isEmpty()) {
+                    impl.setProperties(null);
                 }
                 getTableViewer().refresh(true);
             }
@@ -323,11 +306,11 @@ public class BPMLoggerTable extends Composite implements ICellModifier {
         }
     }
 
-    protected LoggerType1 getTableSelection() {
+    protected PropertyType getTableSelection() {
         if (_propertyTreeTable != null && !_propertyTreeTable.getSelection().isEmpty()) {
             IStructuredSelection ssel = (IStructuredSelection) _propertyTreeTable.getSelection();
-            if (ssel.getFirstElement() instanceof LoggerType1) {
-                return (LoggerType1) ssel.getFirstElement();
+            if (ssel.getFirstElement() instanceof PropertyType) {
+                return (PropertyType) ssel.getFirstElement();
             }
         }
         return null;
@@ -420,16 +403,18 @@ public class BPMLoggerTable extends Composite implements ICellModifier {
      *      java.lang.String)
      */
     public Object getValue(Object element, String property) {
-        if (element instanceof LoggerType1 && property.equalsIgnoreCase(LOG_COLUMN)) {
-            if (((LoggerType1) element).getLog() != null) {
-                return ((LoggerType1) element).getLog();
+        if (element instanceof PropertyType && property.equalsIgnoreCase(NAME_COLUMN)) {
+            if (((PropertyType) element).getName() != null) {
+                return ((PropertyType) element).getName();
             } else {
                 return "";
             }
-        } else if (element instanceof LoggerType1 && property.equalsIgnoreCase(TYPE_COLUMN)) {
-            return new Integer(((LoggerType1) element).getType().getValue());
-        } else if (element instanceof LoggerType1 && property.equalsIgnoreCase(INTERVAL_COLUMN)) {
-            return ((LoggerType1) element).getInterval();
+        } else if (element instanceof PropertyType && property.equalsIgnoreCase(VALUE_COLUMN)) {
+            if (((PropertyType) element).getValue() != null) {
+                return ((PropertyType) element).getValue();
+            } else {
+                return "";
+            }
         }
         return null;
     }
@@ -443,12 +428,12 @@ public class BPMLoggerTable extends Composite implements ICellModifier {
      *      java.lang.String, java.lang.Object)
      */
     public void modify(Object element, String property, final Object value) {
-        if (element instanceof TableItem && property.equalsIgnoreCase(LOG_COLUMN)) {
+        if (element instanceof TableItem && property.equalsIgnoreCase(NAME_COLUMN)) {
             final TableItem ti = (TableItem) element;
             if (getTargetObject() instanceof BPMImplementationType) {
                 final BPMImplementationType impl = (BPMImplementationType) getTargetObject();
-                final LoggerType1 parm = (LoggerType1) ti.getData();
-                if ((value == null && parm.getLog() == null) || (value != null && value.equals(parm.getLog()))) {
+                final PropertyType parm = (PropertyType) ti.getData();
+                if ((value == null && parm.getName() == null) || (value != null && value.equals(parm.getName()))) {
                     return;
                 }
                 if (impl.eContainer() != null) {
@@ -456,58 +441,37 @@ public class BPMLoggerTable extends Composite implements ICellModifier {
                     domain.getCommandStack().execute(new RecordingCommand(domain) {
                         @Override
                         protected void doExecute() {
-                            parm.setLog((String) value);
+                            parm.setName((String) value);
                             getTableViewer().refresh(true);
                         }
                     });
                 } else {
-                    parm.setLog((String) value);
+                    parm.setName((String) value);
                     getTableViewer().refresh(true);
                 }
             }
             fireChangedEvent(this);
             // validate();
-        } else if (element instanceof TableItem && property.equalsIgnoreCase(TYPE_COLUMN)) {
+        } else if (element instanceof TableItem && property.equalsIgnoreCase(VALUE_COLUMN)) {
             final TableItem ti = (TableItem) element;
             if (getTargetObject() instanceof BPMImplementationType) {
                 final BPMImplementationType impl = (BPMImplementationType) getTargetObject();
-                if (impl.eContainer() != null) {
-                    TransactionalEditingDomain domain = SwitchyardSCAEditor.getActiveEditor().getEditingDomain();
-                    domain.getCommandStack().execute(new RecordingCommand(domain) {
-                        @Override
-                        protected void doExecute() {
-                            LoggerType1 parm = (LoggerType1) ti.getData();
-                            LoggerType atype = LoggerType.get(((Integer) value).intValue());
-                            parm.setType(atype);
-                            getTableViewer().refresh(true);
-                        }
-                    });
-                } else {
-                    LoggerType1 parm = (LoggerType1) ti.getData();
-                    LoggerType atype = LoggerType.get(((Integer) value).intValue());
-                    parm.setType(atype);
-                    getTableViewer().refresh(true);
+                final PropertyType parm = (PropertyType) ti.getData();
+                if ((value == null && parm.getValue() == null) || (value != null && value.equals(parm.getValue()))) {
+                    return;
                 }
-            }
-            fireChangedEvent(this);
-            // validate();
-        } else if (element instanceof TableItem && property.equalsIgnoreCase(INTERVAL_COLUMN)) {
-            final TableItem ti = (TableItem) element;
-            if (getTargetObject() instanceof BPMImplementationType) {
-                final BPMImplementationType impl = (BPMImplementationType) getTargetObject();
                 if (impl.eContainer() != null) {
                     TransactionalEditingDomain domain = SwitchyardSCAEditor.getActiveEditor().getEditingDomain();
                     domain.getCommandStack().execute(new RecordingCommand(domain) {
                         @Override
                         protected void doExecute() {
-                            LoggerType1 parm = (LoggerType1) ti.getData();
-                            parm.setInterval((BigInteger) value);
+                            PropertyType parm = (PropertyType) ti.getData();
+                            parm.setValue((String) value);
                             getTableViewer().refresh(true);
                         }
                     });
                 } else {
-                    LoggerType1 parm = (LoggerType1) ti.getData();
-                    parm.setInterval((BigInteger) value);
+                    parm.setValue((String) value);
                     getTableViewer().refresh(true);
                 }
             }
