+++ eclipse/plugins/org.switchyard.tools.ui.bpmn2/src/org/switchyard/tools/ui/bpmn2/component/NewBPMProcessDetailsWizardPage.java
@@ -8,7 +8,7 @@
  * Contributors:
  *     JBoss by Red Hat - Initial implementation.
  ************************************************************************************/
-package org.switchyard.tools.ui.editor.components.bpm;
+package org.switchyard.tools.ui.bpmn2.component;
 
 import java.math.BigInteger;
