+++ eclipse/plugins/org.switchyard.tools.ui.editor/build.properties
@@ -1,4 +1,5 @@
 source.. = src/
+output.. = target/classes/
 bin.includes = META-INF/,\
                .,\
                plugin.xml,\
