+++ eclipse/tests/org.switchyard.tools.ui.tests/META-INF/MANIFEST.MF
@@ -22,6 +22,8 @@ Require-Bundle: org.eclipse.core.runtime;bundle-version="3.4.0",
  org.eclipse.platform,
  javax.servlet,
  javax.xml.rpc,
- org.eclipse.wst.validation
+ org.eclipse.wst.validation,
+ org.switchyard.tools.ui.bpel,
+ org.switchyard.tools.ui.bpmn2
 Bundle-ClassPath: .,
  lib/xmlunit.jar
